Move EnvDteExtensionMethods to Extensions folder
